add del done

+++ lib/services/graphql_service/countries_service/graphql/mutations/country_mutations.dart
@@ -5,8 +5,8 @@ mutation addCountry(
   $capital: String!, 
   $currency: String!, 
   $phone_code: Int!
-  $latitude: float8!, 
-  $longitude: float8!, 
+  $latitude: numeric !, 
+  $longitude: numeric !, 
   ) {
   insert_country_one(object: {name: $name, phone_code: $phone_code, capital: $capital, country_code: $country_code, currency: $currency, longitude: $longitude, latitude: $latitude}) {
     id
@@ -25,7 +25,7 @@ const String updateCountryByPk = r'''
 mutation updateCountry(
   $id: Int!, $name: String!, 
   $capital: String!, $country_code: Int!, 
-  $currency: String!, $latitude: float8!, $longitude: float8!, 
+  $currency: String!, $latitude: numeric!, $longitude: numeric!, 
   $phone_code: Int!) {
   update_country(where: {id: {_eq: $id}}, _set: {name: $name, phone_code: $phone_code, capital: $capital, country_code: $country_code, currency: $currency, longitude: $longitude, latitude: $latitude}) {
     returning {
